refactor(webapp): move the pure streams-version resolver out of the env-bound module
The resolver was written to take its configuration rather than read `env`, but
it still lived beside the singleton that reads it, so the unit test pulled
`env.server` in transitively, which the webapp rules forbid. Same split as
nativeRealtimeClient and its instance module: the pure function has no imports
at all, and the env-bound wrapper stays where the singletons are.

Renames the test to match the function it exercises.
