+++ apps/webapp/app/services/realtime/v1StreamsGlobal.server.ts
@@ -10,6 +10,10 @@ import { singleton } from "~/utils/singleton";
 import type { AuthenticatedEnvironment } from "../apiAuth.server";
 import { RedisRealtimeStreams } from "./redisRealtimeStreams.server";
 import { S2RealtimeStreams } from "./s2realtimeStreams.server";
+import {
+  resolveRealtimeStreamsVersion,
+  type RealtimeStreamsVersionConfig,
+} from "./realtimeStreamsVersion";
 import type { StreamIngestor, StreamResponder } from "./types";
 
 function initializeRedisRealtimeStreams() {
@@ -96,42 +100,7 @@ function streamPrefixFor(environment: AuthenticatedEnvironment, basin: string): 
   return segments.join("/");
 }
 
-export type RealtimeStreamsVersionConfig = {
-  defaultVersion: "v1" | "v2";
-  /** A basin that will actually resolve at read/write time, or undefined if none will. */
-  basin?: string;
-  accessToken?: string;
-  skipAccessTokens: boolean;
-};
-
-/**
- * Resolve the streams version to stamp on a run, falling back to the
- * deployment default when the caller expresses no preference.
- *
- * v2 is only ever returned when S2 can actually serve it. A run stamped v2 on a
- * deployment without S2 is unusable: `getRealtimeStreamInstance` throws for the
- * life of the run, and no read or write against its streams can succeed. v1 is
- * a working backend, so an unsatisfiable v2 degrades to it.
- *
- * The basin must be one that will actually resolve later. Enabling per-org
- * basins is not enough on its own: provisioning is out of band, so an
- * unprovisioned organization has no basin and a global setting may not exist
- * to fall back to.
- */
-export function resolveRealtimeStreamsVersion(
-  streamVersion: string | undefined,
-  config: RealtimeStreamsVersionConfig
-): "v1" | "v2" {
-  const requested = streamVersion ?? config.defaultVersion;
-
-  if (requested !== "v2") {
-    return "v1";
-  }
-
-  const hasCredentials = Boolean(config.accessToken) || config.skipAccessTokens;
-
-  return hasCredentials && Boolean(config.basin) ? "v2" : "v1";
-}
+export type { RealtimeStreamsVersionConfig };
 
 /**
  * Pass `organizationBasinName` wherever the caller has it. It mirrors the
